Fix default checkpoint value and support batch ingestion.
The default values for the memory engine were about the most inefficient its
possible to have with a "dumb" implementation like this.

Refactoring to saner defaults allows an ingestion rate on my test hardware
(SSD based) of 10mb/s of text-form metrics without blowing up memory-usage.

This is tunable with the command line now.